+++ sdks/python/clienthub/api/__init__.py
@@ -8,6 +8,7 @@ from clienthub.api.contacts_api import ContactsApi
 from clienthub.api.health_api import HealthApi
 from clienthub.api.invoices_api import InvoicesApi
 from clienthub.api.lookup_api import LookupApi
+from clienthub.api.marketing_sources_api import MarketingSourcesApi
 from clienthub.api.orders_api import OrdersApi
 from clienthub.api.organizations_api import OrganizationsApi
 from clienthub.api.settings_api import SettingsApi