+++ sdks/typescript/src/apis/index.ts
@@ -7,6 +7,7 @@ export * from './ContactsApi';
 export * from './HealthApi';
 export * from './InvoicesApi';
 export * from './LookupApi';
+export * from './MarketingSourcesApi';
 export * from './OrdersApi';
 export * from './OrganizationsApi';
 export * from './SettingsApi';